Move all the character stuff to a `mod character`.
